Support for tenanted endpoint in Graph explorer.
Currently to go a tenant in Graph Explorer, we need to create a local account in tenant. After this change a tenantId queryString parameter can be passed and then all login requests will go to that tenant.
sample request 
https://developer.microsoft.com/en-us/graph/graph-explorer?tenantId=abc.onmicrosoft.com

if no tenantId is specified, the request goes to common endpoint. 

Tested the redirection is happening correctly.