fix: select only the required values from state and useEffect to get description error message (#3395)
* Update dependencies and workspace

* Select only the required values from the state

Having the selector returning the entire state and then using spread operators to get the values we want will cause components to rerender. This will impact performance because if a component has the entire state and a part of the state it does not use is updated, the component has to rerender. To avoid this, only get what you need from the state for a component.

* Ensure that AutoComplete finishes rendering before calling ValidationProvider

ValidationProvider is updating when AutoComplete tries to call it while rendering.